+++ src/main/java/me/towdium/jecalculation/data/label/ILabel.java
@@ -437,15 +437,15 @@ public interface ILabel {
 
         @Override
         public ILabel increaseAmount() {
-            return setAmount(getAmount() + getMultiplier());
+            int mul = getMultiplier();
+            if (getAmount() == 1 && mul != 1) return setAmount(mul);
+            else return setAmount(getAmount() + mul);
         }
 
         @Override
         public ILabel decreaseAmount() {
-            if (getAmount() <= getMultiplier()) return ILabel.EMPTY;
-            else {
-                return setAmount(getAmount() - getMultiplier());
-            }
+            if (getAmount() <= getMultiplier()) return setAmount(1);
+            else return setAmount(getAmount() - getMultiplier());
         }
 
         @Override