+++ src/main/java/me/towdium/jecalculation/gui/guis/GuiMath.java
@@ -29,7 +29,7 @@ import static me.towdium.jecalculation.data.structure.RecordMath.context;
 @ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
 @OnlyIn(Dist.CLIENT)
-public class GuiMath extends WContainer implements IGui {
+public class GuiMath extends Gui {
     WLcd lcd = new WLcd(7);
     LinkedList<BigDecimal> numbers;
     BigDecimal last;