+++ src/main/java/me/towdium/jecalculation/gui/widgets/WContainer.java
@@ -22,7 +22,6 @@ import java.util.List;
 @OnlyIn(Dist.CLIENT)
 public class WContainer implements IContainer {
     protected List<IWidget> widgets = new ArrayList<>();
-    protected IWidget overlay = null;
 
     public void add(IWidget w) {
         widgets.add(w);
@@ -38,10 +37,6 @@ public class WContainer implements IContainer {
         else widgets.removeAll(Arrays.asList(w));
     }
 
-    public void setOverlay(IWidget overlay) {
-        this.overlay = overlay;
-    }
-
     public void clear() {
         widgets.clear();
     }
@@ -52,75 +47,65 @@ public class WContainer implements IContainer {
 
     @Override
     public boolean onDraw(JecaGui gui, int mouseX, int mouseY) {
-        IWidget[] w = new IWidget[1];
+        Wrapper<IWidget> w = new Wrapper<>(null);
         widgets.forEach(i -> {
-            if (i.onDraw(gui, mouseX, mouseY)) w[0] = i;
+            if (i.onDraw(gui, mouseX, mouseY)) w.value = i;
         });
-        if (w[0] != null) w[0].onDraw(gui, mouseX, mouseY);
-        if (overlay != null) overlay.onDraw(gui, mouseX, mouseY);
+        if (w.value != null) w.value.onDraw(gui, mouseX, mouseY);
         return false;
     }
 
     @Override
     public boolean onMouseClicked(JecaGui gui, int xMouse, int yMouse, int button) {
-        boolean b = overlay != null && overlay.onMouseClicked(gui, xMouse, yMouse, button);
-        return b || new Utilities.ReversedIterator<>(widgets).stream()
+        return new Utilities.ReversedIterator<>(widgets).stream()
                 .anyMatch(i -> i.onMouseClicked(gui, xMouse, yMouse, button));
     }
 
     @Override
     public boolean onKeyPressed(JecaGui gui, int key, int modifier) {
-        boolean b = overlay != null && overlay.onKeyPressed(gui, key, modifier);
-        return b || new Utilities.ReversedIterator<>(widgets).stream()
+        return new Utilities.ReversedIterator<>(widgets).stream()
                 .anyMatch(i -> i.onKeyPressed(gui, key, modifier));
     }
 
     @Override
     public boolean onKeyReleased(JecaGui gui, int key, int modifier) {
-        boolean b = overlay != null && overlay.onKeyReleased(gui, key, modifier);
-        return b || new Utilities.ReversedIterator<>(widgets).stream()
+        return new Utilities.ReversedIterator<>(widgets).stream()
                 .anyMatch(i -> i.onKeyReleased(gui, key, modifier));
     }
 
     @Override
     public boolean onMouseReleased(JecaGui gui, int xMouse, int yMouse, int button) {
-        boolean b = overlay != null && overlay.onMouseReleased(gui, xMouse, yMouse, button);
-        return b || new Utilities.ReversedIterator<>(widgets).stream()
+        return new Utilities.ReversedIterator<>(widgets).stream()
                 .anyMatch(i -> i.onMouseReleased(gui, xMouse, yMouse, button));
     }
 
     @Override
     public boolean onChar(JecaGui gui, char ch, int modifier) {
-        boolean b = overlay != null && overlay.onChar(gui, ch, modifier);
-        return b || new Utilities.ReversedIterator<>(widgets).stream()
+        return new Utilities.ReversedIterator<>(widgets).stream()
                 .anyMatch(i -> i.onChar(gui, ch, modifier));
     }
 
     @Override
     public boolean onMouseScroll(JecaGui gui, int xMouse, int yMouse, int diff) {
-        boolean b = overlay != null && overlay.onMouseScroll(gui, xMouse, yMouse, diff);
-        return b || new Utilities.ReversedIterator<>(widgets).stream()
+        return new Utilities.ReversedIterator<>(widgets).stream()
                 .anyMatch(i -> i.onMouseScroll(gui, xMouse, yMouse, diff));
     }
 
     @Override
     public boolean onMouseDragged(JecaGui gui, int xMouse, int yMouse, int xDrag, int yDrag) {
-        boolean b = overlay != null && overlay.onMouseDragged(gui, xMouse, yMouse, xDrag, yDrag);
-        return b || new Utilities.ReversedIterator<>(widgets).stream()
+        return new Utilities.ReversedIterator<>(widgets).stream()
                 .anyMatch(i -> i.onMouseDragged(gui, xMouse, yMouse, xDrag, yDrag));
     }
 
     @Override
     public boolean onTooltip(JecaGui gui, int xMouse, int yMouse, List<String> tooltip) {
-        boolean b = overlay != null && overlay.onTooltip(gui, xMouse, yMouse, tooltip);
-        return b || new Utilities.ReversedIterator<>(widgets).stream()
+        return new Utilities.ReversedIterator<>(widgets).stream()
                 .anyMatch(i -> i.onTooltip(gui, xMouse, yMouse, tooltip));
     }
 
     @Override
     public boolean getLabelUnderMouse(int xMouse, int yMouse, Wrapper<ILabel> label) {
-        boolean b = overlay != null && overlay.getLabelUnderMouse(xMouse, yMouse, label);
-        return b || new Utilities.ReversedIterator<>(widgets).stream()
+        return new Utilities.ReversedIterator<>(widgets).stream()
                 .anyMatch(i -> i.getLabelUnderMouse(xMouse, yMouse, label));
     }
 }